% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partici_idx.R
\name{partici_idx}
\alias{partici_idx}
\title{Compute Participation Index for Forward and Backward Linkage}
\usage{
partici_idx(wwz, linkage = "forward")
}
\arguments{
\item{wwz}{a decompr class object from the \code{decompr} package.}

\item{linkage}{calculating selecting, \code{forward} or \code{backward}. Wang et al. (2017) puts forward
algorithms of GVC participation index for forward and backward linkage.}
}
\description{
Compute Participation Index for Forward and Backward Linkage
}
\examples{
# load example data
library(decompr)
data(leather)
# create intermediate object (class decompr)
decompr_object <- load_tables_vectors(inter,
                                     final,
                                     countries,
                                     industries,
                                     out)
partici_idx(decompr_object, 'forward')
}
\references{
Wang, Z., et al., MEASURES OF PARTICIPATION IN GLOBAL VALUE CHAINS AND GLOBAL BUSINESS CYCLES. 2017, NBER.
}
